Search for artists and releases
Rate releases inline or select an artist to see releases for that
artist.

At the moment only 25 results are shown for each, but this can be will
be improved for a richer experience soon.

This also fixes the Login for local development. Unfortunately,
MusicBrainz doesn't give a refresh token for localhost so you'll have
login again to get the refresh token every hour when testing.

Fixed #17